moved docker_registry fixture outside of conftest do not check outputs when starting a pipeline as these will be updated set docker registry env variables for sidecar
